Last WIP changes for object store
