+++ .llm/thoughts.md
@@ -1,0 +1,34 @@
+# LLM Thoughts
+
+## Implementing Presigned Download URL Feature
+
+I've implemented a new method in the ObjectStore class to generate presigned URLs for downloading objects from MinIO:
+
+```python
+async def generate_presigned_download_url(
+    self, ref: ObjectRef, expires_in_seconds: int = 3600
+) -> str:
+```
+
+### Design choices:
+
+1. **Input parameter**: The method takes an `ObjectRef` as input since that's the reference system used throughout the codebase.
+
+2. **Expiration time**: Included a configurable expiration time with a reasonable default of 1 hour (3600 seconds).
+
+3. **Validation**: The method validates that the object exists before generating the URL using a HEAD request, which is more efficient than a full GET request.
+
+4. **S3 compatibility**: Used the standard S3 client's `generate_presigned_url` method which should work for both MinIO and AWS S3.
+
+5. **Logging**: Added appropriate logging to track URL generation.
+
+### How this will work with a FastAPI route:
+
+A FastAPI route handler would:
+
+1. Obtain the ObjectRef for the requested file
+2. Call this method to generate the presigned URL
+3. Return the URL in the response
+4. The frontend can then use this URL to download the file directly
+
+The benefit of this approach is that the file download happens directly between the browser and MinIO, without going through the application server, which is more efficient for large files.
