+++ tracecat/api/app.py
@@ -33,7 +33,8 @@ from tracecat.contexts import ctx_role
 from tracecat.db.dependencies import AsyncDBSession
 from tracecat.db.engine import get_async_session_context_manager
 from tracecat.editor.router import router as editor_router
-from tracecat.ee.store.service import setup_store
+from tracecat.ee.store.object_store import setup_store
+from tracecat.ee.store.router import router as object_store_router
 from tracecat.logger import logger
 from tracecat.middleware import RequestLoggingMiddleware
 from tracecat.middleware.security import SecurityHeadersMiddleware
@@ -238,6 +239,11 @@ def create_app(**kwargs) -> FastAPI:
             prefix="/auth",
             tags=["auth"],
         )
+
+    # EE
+    if config.TRACECAT__USE_OBJECT_STORE:
+        logger.info("Including object store router")
+        app.include_router(object_store_router)
 
     # Exception handlers
     app.add_exception_handler(Exception, generic_exception_handler)
