+++ tracecat/dsl/worker.py
@@ -18,7 +18,7 @@ with workflow.unsafe.imports_passed_through():
     from tracecat.dsl.interceptor import SentryInterceptor
     from tracecat.dsl.validation import validate_trigger_inputs_activity
     from tracecat.dsl.workflow import DSLWorkflow
-    from tracecat.ee.store.service import ObjectStore
+    from tracecat.ee.store import service as store_service
     from tracecat.logger import logger
     from tracecat.service import get_activities
     from tracecat.workflow.management.definitions import (
@@ -61,7 +61,7 @@ def all_activities() -> list[Callable]:
         *WorkflowSchedulesService.get_activities(),
         validate_trigger_inputs_activity,
         *WorkflowsManagementService.get_activities(),
-        *get_activities(ObjectStore),
+        *get_activities(store_service),
     ]
